Moved TestReceiverWithNetwork to separate file
