+++ tests/unit/receiver/CMakeLists.txt
@@ -2,6 +2,7 @@
 
 set(TEST_RECEIVER_SOURCES
   test_receiver.cpp
+  test_receiver_with_network.cpp
   main.cpp
 
   ${SACN_SRC}/sacn/receiver.c
